% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disinfo_functions.R
\name{disinfo}
\alias{disinfo}
\title{Create new disinfo object}
\usage{
disinfo(
  claims = data.frame(),
  reviews = data.frame(),
  authors = data.frame(),
  creative_works = data.frame(),
  issues = data.frame(),
  keywords = data.frame(),
  countries = data.frame(),
  languages = data.frame()
)
}
\arguments{
\item{claims}{A claims data frame.}

\item{reviews}{A reviews data frame.}

\item{authors}{A authors or organizations data frame.}

\item{creative_works}{A creative work data frame. Can also be a news article or media object data frame.}

\item{issues}{An issues data frame.}

\item{keywords}{A keywords data frame.}

\item{countries}{A countries data frame.}

\item{languages}{A languages data frame.}
}
\description{
Creates a disinfo object that contains the collection of data frames obtained from the EUvsDisinfo API.
}
\examples{
d <- disinfo()
\dontrun{
d \%>\%
   add_claims("all")

d \%>\%
   add_all("all")
   }
}
