% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disinfo_functions.R
\name{flatten_disinfo}
\alias{flatten_disinfo}
\title{Flatten disinfo object}
\usage{
flatten_disinfo(disinfo)
}
\arguments{
\item{disinfo}{A disinfo object.}
}
\description{
Flatten disinfo object
}
\examples{
d <- disinfo()
\dontrun{
d \%>\%
  add_all("all") \%>\%
  flatten_disinfo()
  }
}
