% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{get_claims}
\alias{get_claims}
\alias{get_claim_reviews}
\title{Get claims and claim reviews}
\usage{
get_claims(pages = 1, published_since = NULL)

get_claim_reviews(
  pages = 1,
  clean_html = TRUE,
  published_since = NULL,
  reviewed_since = NULL
)
}
\arguments{
\item{pages}{Either the number of pages to download or "all". Defaults to 1.}

\item{published_since}{Date string. Only retrieve claims that were published after this date.}

\item{clean_html}{If TRUE, then add another column `text` which is a plain text version of `html_text`.
Defaults to TRUE.}

\item{reviewed_since}{Date string. Only retrieve claim reviews where the claim was reviewed after this date.}
}
\description{
Retrieve the claims and claim reviews.
}
\section{Functions}{
\itemize{
\item \code{get_claim_reviews}: Retrieves the claim reviews which contains the summarized claim and disproof.
}}

\examples{
\dontrun{
get_claims(2)

get_claim_reviews("all")

library(lubridate)
get_claims(2, since = today() - months(3) )
}
}
