% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euvsdisinfo_api_functions.R
\name{get_countries}
\alias{get_countries}
\alias{get_keywords}
\alias{get_languages}
\title{Get countries, keywords and languages}
\usage{
get_countries(pages = 1)

get_keywords(pages = 1)

get_languages(pages = 1)
}
\arguments{
\item{pages}{Either the number of pages to download or "all". Defaults to 1.}
}
\description{
These return lookup tables to the country, keyword and language IDs.
The language IDs are used by the creative work, news articles, and media objects,
whereas the claims data uses keyword and country IDs.
}
\section{Functions}{
\itemize{
\item \code{get_keywords}: Get keywords.

\item \code{get_languages}: Get countries.
}}

\examples{
\dontrun{
get_languages(pages=2)
get_keywords("all")
}
}
