% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{get_creative_works}
\alias{get_creative_works}
\alias{get_news_articles}
\alias{get_media_objects}
\title{Get creative works}
\usage{
get_creative_works(pages = 1, claims_list = NULL)

get_news_articles(pages = 1, claims_list = NULL)

get_media_objects(pages = 1, claims_list = NULL)
}
\arguments{
\item{pages}{Either the number of pages to download or "all". Defaults to 1.}

\item{claims_list}{Vector of claim IDs (of format "/claims/xx") for
which to download the creative work items.}
}
\description{
Returns the creative works (containing both news articles and media objects) or
only news articles or only media objects (videos).
Contains a link to the original source and an archived version, as well as
an abstract (if applicable).
}
\section{Functions}{
\itemize{
\item \code{get_news_articles}: Get only news articles.

\item \code{get_media_objects}: Get only media objects (videos).
}}

\examples{
\dontrun{
get_news_articles(pages=1)
get_media_objects("all")
get_creative_works(claims_list = c("/claims/36", "/claims/24"))
}
}
