% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{get_issues}
\alias{get_issues}
\title{Get issues}
\usage{
get_issues(pages = 1)
}
\arguments{
\item{pages}{Either the number of pages to download or "all". Defaults to 1.}
}
\description{
Returns the weekly issues that are sent by [EU vs Disinfo](https://euvsdisinfo.eu/)
in their newsletter.
The issues link to the claim IDs contained in the issue.
}
\examples{
\dontrun{
get_issues(pages=1)
}
}
