% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euvsdisinfo_api_functions.R
\name{get_organizations}
\alias{get_organizations}
\alias{get_authors}
\title{Get authors and organizations}
\usage{
get_organizations(pages = 1)

get_authors(pages = 1)
}
\arguments{
\item{pages}{Either the number of pages to download or "all". Defaults to 1.}
}
\description{
Currently, there is no difference between organizations and authors and
both download the same data regarding the outlet organization.
}
\section{Functions}{
\itemize{
\item \code{get_authors}: Get organizations
}}

\examples{
\dontrun{
get_authors(pages=2)
get_organizations("all")
}
}
