% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disinfo_functions.R
\name{add_claims}
\alias{add_claims}
\alias{add_reviews}
\alias{add_authors}
\alias{add_creative_works}
\alias{add_issues}
\alias{add_keywords}
\alias{add_countries}
\alias{add_languages}
\alias{add_news_articles}
\alias{add_media_objects}
\alias{add_all}
\title{Add data to disinfo object}
\usage{
add_claims(disinfo, pages = 1, published_since = NULL)

add_reviews(
  disinfo,
  pages = 1,
  clean_html = TRUE,
  published_since = NULL,
  reviewed_since = NULL
)

add_authors(disinfo, pages = 1)

add_creative_works(disinfo, pages = 1)

add_issues(disinfo, pages = 1)

add_keywords(disinfo, pages = 1)

add_countries(disinfo, pages = 1)

add_languages(disinfo, pages = 1)

add_news_articles(disinfo, pages = 1)

add_media_objects(disinfo, pages = 1)

add_all(disinfo, pages = 1, reviewed_since = NULL, published_since = NULL)
}
\arguments{
\item{disinfo}{A disinfo object.}

\item{pages}{Either the number of pages to download or "all". Defaults to 1.}

\item{published_since}{Date string. Only retrieve claims or claim reviews where the claims were published after this date.
It is currently only possible to restrict the claims and claim reviews by the publishing date.}

\item{clean_html}{If TRUE, then add another column `text` which is a plain text version of `html_text`.
Defaults to TRUE.}

\item{reviewed_since}{Date string. Only retrieve claim reviews that were reviewed after this date.
It is currently only possible to restrict the claim reviews by the reviewing date.}
}
\description{
Add the different data tables to the disinfo object by calling the EUvsDisinfo API.
}
\section{Functions}{
\itemize{
\item \code{add_reviews}: Download claim reviews data and add to disinfo object.

\item \code{add_authors}: Download authors data and add to disinfo object.

\item \code{add_creative_works}: Download creative works data and add to disinfo object.
If either claims or claim reviews already exists in disinfo, then it will only
download creative works that link to these claims/reviews. (Only works if there are less than 50 claim IDs)

\item \code{add_issues}: Download issues data and add to disinfo object.

\item \code{add_keywords}: Download keywords data and add to disinfo object.

\item \code{add_countries}: Download countries data and add to disinfo object.

\item \code{add_languages}: Download languages data and add to disinfo object.

\item \code{add_news_articles}: Download news articles data and add to disinfo object. Will be
written to `creative_works` and overwrites any previous creative works data.
If you want to have both news articles and media objects, use `add_creative_works()`.
If either claims or claim reviews already exists in disinfo, then it will try to only
download news articles that link to these claims/reviews. (Only works if there are less than 50 claim IDs)

\item \code{add_media_objects}: Download media objects data and add to disinfo object. Will be
written to `creative_works` and overwrites any previous creative works data.
If you want to have both news articles and media objects, use `add_creative_works()`.
If either claims or claim reviews already exists in disinfo, then it will only
download media objects that link to these claims/reviews. (Only works if there are less than 50 claim IDs)

\item \code{add_all}: Downloads all 8 data tables from the API and adds them to the disinfo object.
}}

\examples{
d <- disinfo()
\dontrun{
d \%>\%
   add_claims()

d \%>\%
    add_claims(published_since = today() - months(3) ) \%>\%
    add_reviews(published_since = today() - months(3) )

d \%>\%
   add_all()
}
}
