% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euvsdisinfo_api_functions.R
\name{get_claims}
\alias{get_claims}
\alias{get_claim_reviews}
\title{Get claims and claim reviews}
\usage{
get_claims(pages = 1)

get_claim_reviews(pages = 1)
}
\arguments{
\item{pages}{Either the number of pages to download or "all". Defaults to 1.}
}
\description{
Returns the claims (containing meta-data regarding the claims) a
and claim reviews (containing the summarized claim and disproof).
}
\section{Functions}{
\itemize{
\item \code{get_claim_reviews}: Get claim reviews.
}}

\examples{
\dontrun{
get_claims(2)
}
\dontrun{
get_claim_reviews("all")
}
}
